ref(js): Rewrite ThemeAndStyleProvider as a functional component
Theme does not need to be kept in state, since we can just rely on the
provided config object being updated from the withConfig
